+++ closed_session.go
@@ -79,9 +79,8 @@ func (s *closedLocalSession) handlePacketImpl(_ *receivedPacket) {
 	}
 }
 
-func (s *closedLocalSession) Close() error {
+func (s *closedLocalSession) shutdown() {
 	s.destroy(nil)
-	return nil
 }
 
 func (s *closedLocalSession) destroy(error) {
@@ -108,6 +107,6 @@ func newClosedRemoteSession(pers protocol.Perspective) packetHandler {
 }
 
 func (s *closedRemoteSession) handlePacket(*receivedPacket)         {}
-func (s *closedRemoteSession) Close() error                         { return nil }
+func (s *closedRemoteSession) shutdown()                            {}
 func (s *closedRemoteSession) destroy(error)                        {}
 func (s *closedRemoteSession) getPerspective() protocol.Perspective { return s.perspective }