+++ integrationtests/self/rtt_test.go
@@ -72,7 +72,7 @@ var _ = Describe("non-zero RTT", func() {
 					data, err := ioutil.ReadAll(str)
 					Expect(err).ToNot(HaveOccurred())
 					Expect(data).To(Equal(PRData))
-					sess.Close()
+					sess.CloseWithError(0, "")
 					Eventually(done).Should(BeClosed())
 				})
 			}