+++ integrationtests/self/uni_stream_test.go
@@ -75,7 +75,7 @@ var _ = Describe("Unidirectional Streams", func() {
 			sess, err := server.Accept(context.Background())
 			Expect(err).ToNot(HaveOccurred())
 			runReceivingPeer(sess)
-			sess.Close()
+			sess.CloseWithError(0, "")
 		}()
 
 		client, err := quic.DialAddr(